+++ stocks/management/commands/import_stock_trades_utilities/DataTypeValidator.py
@@ -1,0 +1,7 @@
+class DataTypeValidator:
+    @staticmethod
+    def validate_data_type(variable, data_type, exception_error_message):
+        if isinstance(variable, data_type):
+            return variable
+        else:
+            raise ValueError(exception_error_message)
